Extract gradient modifier
